working on getting tests to work
